NETS/requirements.txt
torch==1.6.0
pandas==1.1.3
scipy==1.5.2
tqdm==4.51.0
matplotlib==3.3.2
numpy==1.19.2
PyYAML==5.4.1
scikit_learn==0.24.1
